+++ test/bones_api_test.reflection.g.dart
@@ -61,7 +61,7 @@ class MyInfoModule$reflection extends ClassReflection<MyInfoModule>
   }
 
   @override
-  Version get languageVersion => Version.parse('3.3.0');
+  Version get languageVersion => Version.parse('3.4.0');
 
   @override
   MyInfoModule$reflection withObject([MyInfoModule? obj]) =>